Igushik River Escapement and Age Composition Data

A data frame containing annual escapement (S), total run (N), and age composition columns (A3-A8) for the Igushik River. These observations are typically used to generate brood tables, estimate recruitment, and prepare inputs for stock-recruit analyses.

data_Igushik

Format

A data frame with one row per year and the following variables:

yr

Calendar year of observation.

S

Spawning escapement (numeric).

N

Total return (numeric).

A3

Count or proportion of age-3 returning fish.

A4

Count or proportion of age-4 returning fish.

A5

Count or proportion of age-5 returning fish.

A6

Count or proportion of age-6 returning fish.

A7

Count or proportion of age-7 returning fish.

A8

Count or proportion of age-8 returning fish.

Source

Internal ADF&G escapement and age composition data, collected at the Igushik River.

Details

Age composition fields may represent either expanded counts or proportions, depending on how the dataset was processed prior to being archived. These values are used by functions such as get_age() and get_brood(). The dataset originates from Igushik River escapement goal analyses.

Examples

head(data_Igushik)
#>     yr      S      N  A3     A4     A5    A6 A7 A8
#> 1 1963  92184 136351   0  75779  51359  9213  0  0
#> 2 1964 128532 191054   0  41912 135654 13488  0  0
#> 3 1965 180840 297253   0  28452 252608 16194  0  0
#> 4 1966 206360 310078   0  17671 262777 29630  0  0
#> 5 1967 281772 412374 105 196850 205905  9514  0  0
#> 6 1968 194508 290686 299 125081 159764  5541  0  0